Implementacion de la RNF9, en donde se manejan las excpeciones para que no ocurra ningun fallo a la hora de ejecutar el programa. Tanto excepciones para manipular el controlador de la vista, como excepciones de persistencia y logica.

+++ binarytree/src/main/java/co/edu/uptc/application/service/EliminarArbolService.java
@@ -3,6 +3,7 @@ package co.edu.uptc.application.service;
 import java.util.List;
 
 import co.edu.uptc.domain.exception.NoTreeSelectedException;
+import co.edu.uptc.domain.exception.TreeNotFoundException;
 import co.edu.uptc.domain.repository.BinaryTreeRepository;
 
 public class EliminarArbolService {
@@ -15,6 +16,10 @@ public class EliminarArbolService {
     public void eliminarArbol(String nombre) {
         if (nombre == null || !repository.exists(nombre)) {
             throw new NoTreeSelectedException();
+        }
+
+        if(!repository.exists(nombre)){
+            throw new TreeNotFoundException(nombre);
         }
         repository.delete(nombre);
     }
